fix(convex): rename bunny-urls.ts to bunnyUrls.ts for valid module path
Convex rejette les tirets dans les chemins de modules (deploy:
"InvalidConfig: lib/bunny-urls.js is not a valid path"). Renomme en
camelCase pour suivre la convention existante (lib/batchFetch.ts) et
met a jour l'import + le commentaire dans lib/bunny.ts. convex-test
ne validant pas cette regle, l'echec n'apparaissait qu'au deploiement.
